+++ ChangeLog.md
@@ -1,3 +1,6 @@
+## [UNRELEASED]
+- Re-connect when losing connection to etcd
+
 ## [1.0.0] 2024-10-15
 - Added basic policy and authentication support
 - Finished deletion support