+++ ChangeLog.md
@@ -22,6 +22,7 @@
 - Implement special characters: `${*}`, `${?}`, `${$}`
 - Support variable on resource policy statement
 - Prevent services from being destroyed while etcd is not booted up
+- Restrictions on naming buckets have been resolved (e.g. cannot start with a number)
 
 ## [1.0.0] 2024-10-15
 - Added basic policy and authentication support